requirements.txt
streamlit
pandas
plotly
kaleido
fpdf2